layout/style/test/test_bug732153.html

<!doctype html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=732153
-->
<title>Test for Bug 732153</title>
<script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=732153">Mozilla Bug 732153</a>
<div></div>
<script>
var div = document.querySelector("div");
[
  "",
  "backface-visibility: hidden",
  "transform-style: preserve-3d",
  "backface-visibility: hidden; transform-style: preserve-3d",
].forEach(function(style) {
  div.setAttribute("style", style);
  is(getComputedStyle(div).transform, "none",
    "Computed 'transform' with style=\"" + style + '"');
});
</script>
